ultimos cambios, sigue dando null el costo estimado, cuando haces el post a solicitudes

+++ microservicio-viajes/src/main/java/com/proyectoviajes/viajems/dto/CalculoEstimadoResponse.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class CalculoEstimadoResponse {
     private Double distanciaTotalKm;
-    private Long tiempoEstimadoHoras;
+    private Double tiempoEstimadoHoras;
     private Double costoEstimado;
 }
